refactor: converge extractor and renderer modules
