--- pdf_craft/config.py
@@ -1,2 +1,0 @@
-MAX_TITLE_CV = 0.025
-MAX_LEVELS = 4
