--- pdf_craft/epub/__init__.py
@@ -1,1 +1,0 @@
-from .render import render_epub_file
