+++ pdf_craft/extractor/toc/__init__.py
@@ -1,1 +1,2 @@
-"""TOC analysis namespace."""
+from .analysing import analyse_toc
+from .types import Toc, TocInfo, decode, encode, iter_toc
