+++ pdf_craft/extractor/toc/types.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Generator
 from xml.etree.ElementTree import Element
 
-from ..common import indent
+from ...common import indent
 
 
 @dataclass
