--- pdf_craft/sequence/__init__.py
@@ -1,19 +1,0 @@
-from .chapter import (
-    AssetLayout,
-    AssetRef,
-    BlockLayout,
-    BlockMember,
-    Chapter,
-    InlineExpression,
-    ParagraphLayout,
-    Reference,
-    RefIdMap,
-    decode,
-    encode,
-    references_to_map,
-    search_references_in_chapter,
-)
-from .content import Content
-from .generation import generate_chapter_files
-from .mark import Mark, NumberClass, NumberStyle
-from .reader import create_chapters_reader
