--- pdf_craft/toc/__init__.py
@@ -1,2 +1,0 @@
-from .analysing import analyse_toc
-from .types import Toc, TocInfo, decode, encode, iter_toc
