+++ tests/test_toc_text.py
@@ -1,6 +1,6 @@
 import unittest
 
-from pdf_craft.toc.text import normalize_text
+from pdf_craft.extractor.toc.text import normalize_text
 
 
 class TestNormalizeText(unittest.TestCase):
